refactor(test): extract Upload v2 fixture lease lifecycle
Design: Spec 008

Behavior: no finding closed; behavior-preserving lifecycle extraction.

Characterization: npm test --prefix tests/e2e -- gfsUploadV2Fixtures.test.ts passed before and after extraction with 13 tests.
