+++ tests/e2e/gfsUploadV2Fixtures.ts
@@ -1,8 +1,15 @@
 import { createHash } from 'node:crypto'
 import { constants, createReadStream } from 'node:fs'
-import { type FileHandle, mkdtemp, open } from 'node:fs/promises'
+import { mkdtemp, open } from 'node:fs/promises'
 import os from 'node:os'
 import path from 'node:path'
+import {
+  disposeFixtureLeaseForTest,
+  registerFixtureLeaseForTest,
+  setFixtureHandleForTest,
+} from './gfsUploadV2FixtureLease.js'
+
+export { setFixtureHandleForTest } from './gfsUploadV2FixtureLease.js'
 
 /** Test-only default for runtime Upload v2 journeys; GFSC remains the product-policy authority. */
 export const E2E_GFS_UPLOAD_V2_DEFAULT_PRODUCT_MAX_BYTES = 209_715_200
@@ -22,36 +29,13 @@ export interface DiskUploadFixture {
   sha256: string
 }
 
-interface FixtureHandle {
-  truncate(length: number): Promise<void>
-  close(): Promise<void>
-}
-
-interface FixtureLease {
-  handle: FixtureHandle
-  state: 'active' | 'disposed' | 'neutralize-failed-closed' | 'close-failed'
-  neutralized: boolean
-  failure?: unknown
-}
-
-const fixtureLeases = new WeakMap<DiskUploadFixture, FixtureLease>()
 let afterFixtureHandleForTest: ((fixture: DiskUploadFixture) => Promise<void> | void) | undefined
-let fixtureHandleForTest:
-  | ((handle: FileHandle, fixture: DiskUploadFixture) => FixtureHandle)
-  | undefined
 
 /** Test-only seam for exercising cleanup after a failure that follows handle acquisition. */
 export function setAfterFixtureHandleForTest(
   hook: ((fixture: DiskUploadFixture) => Promise<void> | void) | undefined
 ): void {
   afterFixtureHandleForTest = hook
-}
-
-/** Test-only seam for exercising retained-handle disposal failures. */
-export function setFixtureHandleForTest(
-  hook: ((handle: FileHandle, fixture: DiskUploadFixture) => FixtureHandle) | undefined
-): void {
-  fixtureHandleForTest = hook
 }
 
 /**
@@ -114,12 +98,7 @@ async function createFixture(
       byteLength,
       sha256: '',
     }
-    const ownedHandle = fixtureHandleForTest?.(handle, fixture) ?? handle
-    fixtureLeases.set(fixture, {
-      handle: ownedHandle,
-      state: 'active',
-      neutralized: false,
-    })
+    const ownedHandle = registerFixtureLeaseForTest(fixture, handle)
     await afterFixtureHandleForTest?.(fixture)
     await ownedHandle.truncate(byteLength)
     fixture.sha256 = await sha256File(filePath)
@@ -127,7 +106,7 @@ async function createFixture(
   } catch (error) {
     if (fixture) {
       try {
-        await disposeFixtureLease(fixture)
+        await disposeFixtureLeaseForTest(fixture)
       } catch (cleanupError) {
         throw new AggregateError(
           [error, cleanupError],
@@ -146,86 +125,5 @@ export async function sha256File(filePath: string): Promise<string> {
 }
 
 export async function removeDiskUploadFixture(fixture: DiskUploadFixture): Promise<void> {
-  const lease = fixtureLeases.get(fixture)
-  if (!lease) {
-    throw new Error(`refusing to neutralize an unowned GFS v2 fixture: ${fixture.filePath}`)
-  }
-  await disposeFixtureLease(fixture)
-}
-
-async function disposeFixtureLease(fixture: DiskUploadFixture): Promise<void> {
-  const lease = fixtureLeases.get(fixture)
-  if (!lease) {
-    throw new Error(`refusing to neutralize an unowned GFS v2 fixture: ${fixture.filePath}`)
-  }
-  if (lease.state === 'disposed') return
-  if (lease.state === 'neutralize-failed-closed') {
-    throw (
-      lease.failure ??
-      new Error(`GFS v2 fixture disposal failed without a recorded cause: ${fixture.filePath}`)
-    )
-  }
-  if (lease.state === 'close-failed') {
-    await retryCloseAfterFailure(fixture, lease)
-    return
-  }
-  try {
-    await lease.handle.truncate(0)
-    lease.neutralized = true
-    await lease.handle.close()
-    lease.state = 'disposed'
-  } catch (error) {
-    await closeAfterDisposalFailure(lease, error)
-  }
-}
-
-async function closeAfterDisposalFailure(lease: FixtureLease, failure: unknown): Promise<never> {
-  try {
-    await lease.handle.close()
-  } catch (closeError) {
-    lease.failure = combineFixtureDisposalFailures(lease.neutralized, failure, closeError)
-    lease.state = 'close-failed'
-    throw lease.failure
-  }
-  if (lease.neutralized) {
-    lease.state = 'disposed'
-  } else {
-    lease.failure = failure
-    lease.state = 'neutralize-failed-closed'
-  }
-  throw failure
-}
-
-async function retryCloseAfterFailure(
-  fixture: DiskUploadFixture,
-  lease: FixtureLease
-): Promise<void> {
-  const previousFailure =
-    lease.failure ??
-    new Error(`GFS v2 fixture disposal failed without a recorded cause: ${fixture.filePath}`)
-  try {
-    await lease.handle.close()
-  } catch (closeError) {
-    lease.failure = combineFixtureDisposalFailures(lease.neutralized, previousFailure, closeError)
-    throw lease.failure
-  }
-  if (lease.neutralized) {
-    lease.state = 'disposed'
-    return
-  }
-  lease.state = 'neutralize-failed-closed'
-  throw previousFailure
-}
-
-function combineFixtureDisposalFailures(
-  neutralized: boolean,
-  failure: unknown,
-  closeError: unknown
-): AggregateError {
-  return new AggregateError(
-    [failure, closeError],
-    neutralized
-      ? 'failed to close neutralized GFS v2 fixture'
-      : 'failed to neutralize and close GFS v2 fixture'
-  )
+  await disposeFixtureLeaseForTest(fixture)
 }
